pyproject.toml
[tool.poetry]
name = "nyt-app"
version = "0.1.0"
description = ""
authors = ["Andy Kuny <andrew.kuny@gsa.gov>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
